Add files via upload
Got a polynomial regression line working for the ranged weapon graph and also added a few more weapons to the list.
I'd like to continue adding more weapons to check how accurate the regression is.
I think up next will be adding in a way to change the unit you are trying to wound in the Weapon_Value function. Or at least feed it some numbers and allow the person using the function to play with the graph and values.