Add files via upload
Today I added in some new functionality for the plotting function! You can now check how a member of a unit fairs against the armies ranged weapons, the default is what it was before.

I also did a little bit of cleaning up, replacing some names which didn't fit well with some others which are a little more telling of what's being used or called.

I have started to think about the new edition as well. Since the new edition will be removing and adding stats to members and weapons I will have to consider how to implement them and try and preempt it so that when the new edition comes out I can make all the changes I need to in a day.